rename circuit.rs
