+++ halo2_gadgets/src/sinsemilla.rs
@@ -472,7 +472,7 @@ pub(crate) mod tests {
             NonIdentityPoint, ScalarFixed,
         },
         sinsemilla::primitives::{self as sinsemilla, K},
-        tests::circuit::{
+        tests::test_utils::{
             fixed_verification_key_test_with_circuit, serialized_proof_test_case_with_circuit,
         },
         utilities::lookup_range_check::{LookupRangeCheck, LookupRangeCheckConfig},
