+++ halo2_gadgets/src/sinsemilla/merkle.rs
@@ -184,7 +184,7 @@ pub mod tests {
             tests::{TestCommitDomain, TestHashDomain},
             HashDomains,
         },
-        tests::circuit::{
+        tests::test_utils::{
             fixed_verification_key_test_with_circuit, serialized_proof_test_case_with_circuit,
         },
         utilities::{
