+++ halo2_gadgets/src/tests.rs
@@ -1,1 +1,1 @@
-pub(crate) mod circuit;
+pub(crate) mod test_utils;
